// source/frontend/src/models/api/generated/ApiGen_Concepts_ResearchFileProject.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseConcurrent } from './ApiGen_Base_BaseConcurrent';
import { ApiGen_Concepts_Project } from './ApiGen_Concepts_Project';
import { ApiGen_Concepts_ResearchFile } from './ApiGen_Concepts_ResearchFile';

// LINK: @backend/apimodels/Models/Concepts/ResearchFile/ResearchFileProjectModel.cs
export interface ApiGen_Concepts_ResearchFileProject extends ApiGen_Base_BaseConcurrent {
  id: number;
  project: ApiGen_Concepts_Project | null;
  projectId: number | null;
  file: ApiGen_Concepts_ResearchFile | null;
  fileId: number;
}